reduce: Fix return type of internal utilities in a couple of places

+++ simdpp/detail/insn/i_reduce_or.h
@@ -41,7 +41,7 @@ SIMDPP_INL uint8_t i_reduce_or(const uint8x16& a)
 }
 
 #if SIMDPP_USE_AVX2
-SIMDPP_INL uint16_t i_reduce_or(const uint8x32& a)
+SIMDPP_INL uint8_t i_reduce_or(const uint8x32& a)
 {
     uint8x16 r = detail::extract128<0>(a);
     r = bit_or(r, detail::extract128<1>(a));
